+++ src/include/tlct/convert/patchsize/params.hpp
@@ -22,21 +22,21 @@ public:
     // Initialize from
     [[nodiscard]] static inline PsizeParams_ fromConfigs(const TArrange& arrange, const TCvtConfig& cvt_cfg);
 
-    double pattern_size;
-    double pattern_shift;
+    float pattern_size;
+    float pattern_shift;
     int min_psize;
 };
 
 template <tcfg::concepts::CArrange TArrange>
 PsizeParams_<TArrange> PsizeParams_<TArrange>::fromConfigs(const TArrange& arrange, const TCvtConfig& cvt_cfg)
 {
-    const double pattern_size = arrange.getDiameter() * cvt_cfg.pattern_size;
-    const double radius = arrange.getDiameter() / 2.0;
-    const double half_pattern_size = pattern_size / 2.0;
-    const double max_pattern_shift =
+    const float pattern_size = arrange.getDiameter() * cvt_cfg.pattern_size;
+    const float radius = arrange.getDiameter() / 2.0;
+    const float half_pattern_size = pattern_size / 2.0;
+    const float max_pattern_shift =
         std::sqrt((radius - half_pattern_size) * (radius + half_pattern_size)) - half_pattern_size;
-    const double candidate_pattern_shift = radius * cvt_cfg.max_psize;
-    const double pattern_shift = std::min(max_pattern_shift, candidate_pattern_shift);
+    const float candidate_pattern_shift = radius * cvt_cfg.max_psize;
+    const float pattern_shift = std::min(max_pattern_shift, candidate_pattern_shift);
 
     const int min_psize = (int)std::round(0.75 * pattern_size);
